requirements.txt
streamlit
python-dotenv
crewai
crewai-tools
langchain-community
pydantic
openai